⛑️ test : LastFrame 팩토리 메소드 패턴으로 중복 제거
